Save Commit: Dental Schools update. Added student edit window and support for editing it.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@10895 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/UserGroups.cs
@@ -68,6 +68,12 @@ namespace OpenDentBusiness{
 			if(table.Rows[0][0].ToString()!="0"){
 				throw new Exception(Lans.g("UserGroups","Must move users to another group first."));
 			}
+			if(PrefC.GetLong(PrefName.SecurityGroupForStudents)==group.UserGroupNum) {
+				throw new Exception(Lans.g("UserGroups","Group is the default group for students and cannot be deleted.  Change the default student group before deleting."));
+			}
+			if(PrefC.GetLong(PrefName.SecurityGroupForInstructors)==group.UserGroupNum) {
+				throw new Exception(Lans.g("UserGroups","Group is the default group for instructors and cannot be deleted.  Change the default instructors group before deleting."));
+			}
 			command= "DELETE FROM usergroup WHERE UserGroupNum='"
 				+POut.Long(group.UserGroupNum)+"'";
  			Db.NonQ(command);
